feat: 특정 회원의 코인 거래 내역 조회 구현

+++ src/main/java/ksh/example/mybit/domain/Trade.java
@@ -16,7 +16,7 @@ public class Trade {
     @Column(name = "trade_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private Long executedQuantity;
+    private BigDecimal executedQuantity;
     private BigDecimal executedPrice;
     private Integer executedAmount;
